sdk: refactoring the Azure CLI authorizer into a single file/making the internals private
